Rename regexPattern to regexPatterns
