+++ src/components/LetterProofing.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import NotificationModal from "../login/components/Modals/NotificationModal";
 import ConfirmModal from "../containers/ConfirmModal";
-import { shortCodePattern } from "../login/app_utils/validation/regexPattern";
+import { shortCodePattern } from "../login/app_utils/validation/regexPatterns";
 
 class LetterProofingButton extends Component {
   render() {
