+++ src/components/Security.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import EduIDButton from "components/EduIDButton";
 import NotificationModal from "../login/components/Modals/NotificationModal";
 import ConfirmModal from "../containers/ConfirmModal";
-import { securityKeyPattern } from "../login/app_utils/validation/regexPattern";
+import { securityKeyPattern } from "../login/app_utils/validation/regexPatterns";
 import "../login/styles/index.scss";
 
 class Security extends Component {
