+++ src/login/app_utils/validation/regexPatterns.js
